# api_service/app/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.DataResponse:
    properties:
      count:
        type: integer
      data: {}
    type: object
...
  models.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  models.FlangeDTO:
    properties:
      short:
        type: string
      title:
        type: string
    required:
    - short
    - title
    type: object
  models.IdResponse:
    properties:
      id:
        type: string
      message:
        type: string
    type: object
...
  models.MaterialsDTO:
    properties:
      title:
        type: string
      typeMat:
        type: string
    required:
    - title
    - typeMat
    type: object
...
  models.UpdateMatDTO:
    properties:
      change:
        type: string
      materials:
        items:
          $ref: '#/definitions/proto.AddMaterials'
        type: array
      type:
        type: string
    required:
    - materials
    - type
    type: object
...
  proto.AddMaterials:
    properties:
      short:
        type: string
      title:
        type: string
    type: object
...
  proto.Flange:
    properties:
      id:
        type: string
      short:
        type: string
      title:
        type: string
    type: object
  proto.Materials:
    properties:
      id:
        type: string
      title:
        type: string
      typeMat:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Service for Sealur
  title: Sealur
  version: 0.14.0
paths:
...
  /sealur-pro/additionals/{id}/mat:
    patch:
      consumes:
      - application/json
      description: обновление материалов
      parameters:
      - description: additional materials info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMatDTO'
      - description: addit id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Mat
      tags:
      - Sealur Pro -> additionals
...
  /sealur-pro/flanges:
    get:
      consumes:
      - application/json
      description: получение всех стандартов на фланцы
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.Flange'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Flanges
      tags:
      - Sealur Pro -> flanges
    post:
      consumes:
      - application/json
      description: создание стандарта на фланцы
      parameters:
      - description: flange info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.FlangeDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Flange
      tags:
      - Sealur Pro -> flanges
  /sealur-pro/flanges/{id}:
    delete:
      consumes:
      - application/json
      description: удаление стандарта на фланцы
      parameters:
      - description: flange id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Flange
      tags:
      - Sealur Pro -> flanges
    put:
      consumes:
      - application/json
      description: обновление стандарта на фланцы
      parameters:
      - description: flange info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.FlangeDTO'
      - description: flange id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Flange
      tags:
      - Sealur Pro -> flanges
...
  /sealur-pro/materials:
    get:
      consumes:
      - application/json
      description: получение материалов (для опроса)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.DataResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/proto.Materials'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Materials
      tags:
      - Sealur Pro -> materials
    post:
      consumes:
      - application/json
      description: создание материала (для опроса)
      parameters:
      - description: materials info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.MaterialsDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Materials
      tags:
      - Sealur Pro -> materials
  /sealur-pro/materials/{id}:
    delete:
      consumes:
      - application/json
      description: удаление материала (для опроса)
      parameters:
      - description: material id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Materials
      tags:
      - Sealur Pro -> materials
    put:
      consumes:
      - application/json
      description: обновление материала (для опроса)
      parameters:
      - description: material info
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.MaterialsDTO'
      - description: material id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Materials
      tags:
      - Sealur Pro -> materials
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"